+++ src/lib/components/types/Student.ts
@@ -1,0 +1,10 @@
+type Student = {
+  name: string
+  email: string
+  secondaryEmail: string
+  phone: string
+  grade: number
+  school: string
+}
+
+export default Student;
